feat: changes promoted events button behavior
